Numerous Bugfixes.
Increased streamlining and validation for using the 'xtDataCell', 'ppDataCell', and 'sdoMultiMat' classes. 

Consolidated and depreciated some of the older direct function calls which were too unwieldy to be used practically. 

Reduced redundancy in the estimation of the SDO. 

Added modularity to the error prediction so that further tests may be added easily. 

Added a dedicated 'external' folder for outside support functions. 

Modularized the SDO estimation algorithms and model-hypothesis matrix generation scripts so that future variants can be easily incoporated. 

Added the 'predictionError' class, containing the distance between hypothesis-predicted and observed post-spike probability distributions, for better handling and stability. 

Added support for the parallel compute toolbox, allowing for batch processing of multiple SDO s on GPU. 

Added the 'asymmetric' SDO estimation algorithm which better separates the drift vs. the diffusion components, and better ellucidates matrix motifs. 

Generated a new colormap to dynamically rescale with the SDO matrices, so that the background is always 'white' with the off-diagonals (positive) as red/yellow (better resembling the paper). 

Updated the sdoAnalysis_demo.mlx live script with the newer methods.